+++ indico/queries/workflow.py
@@ -5,8 +5,7 @@ from typing import List, Union, Dict
 from indico.client.request import GraphQLRequest, RequestChain, Debouncer
 from indico.errors import IndicoError, IndicoInputError
 from indico.queries.storage import UploadDocument
-from indico.types import Job, Submission, Workflow, SUBMISSION_RESULT_VERSIONS, ModelTaskType, \
-    NewLabelsetArguments, NewQuestionaireArguments
+from indico.types import Job, Submission, Workflow, SUBMISSION_RESULT_VERSIONS
 from indico.types.utils import cc_to_snake, snake_to_cc
 
 
@@ -499,120 +498,3 @@ class CreateWorkflow(GraphQLRequest):
             **super().process_response(response)["createWorkflow"]["workflow"]
         )
 
-
-class AddModelGroupComponent(GraphQLRequest):
-    """
-            Adds extraction or classification component to a workflow.
-            Available on 5.0+.
-            Returns workflow with updated component list.
-    Args:
-         workflow_id: int,
-         dataset_id: int,
-         name: str,
-        source_column_id: int,
-        after_component_id: int,
-        labelset_column_id: int = None
-        new_labelset_args: NewLabelsetArguments = None
-        new_questionnaire_args: NewQuestionaireArguments = None
-
-    """
-    query = """
-            mutation addModelGroup(
-          $workflowId: Int!, 
-          $name: String!, 
-          $datasetId: Int!, 
-          $sourceColumnId: Int!, 
-          $afterComponentId: Int, 
-          $labelsetColumnId: Int,
-          $newLabelsetArgs: NewLabelsetInput,
-          $questionnaireArgs: QuestionnaireInput,
-          $modelTrainingOptions: JSONString
-        ) {
-          addModelGroupComponent(workflowId: $workflowId, name: $name, datasetId: $datasetId, 
-          sourceColumnId: $sourceColumnId, afterComponentId: $afterComponentId, labelsetColumnId: $labelsetColumnId,
-          modelTrainingOptions: $modelTrainingOptions,
-          
-    newLabelsetArgs: $newLabelsetArgs,
-    questionnaireArgs: $questionnaireArgs,) {
-            workflow {
-                id
-                components {
-                                id
-                                componentType
-                                reviewable
-                                filteredClasses
-                                ... on ModelGroupComponent {
-                                    taskType
-                                    modelType
-                                    modelGroup {
-                                        status
-                                      id
-                                      name
-                                      taskType
-                                      selectedModel{
-                                        id
-                                      }
-                                    }
-                                }
-
-                            }
-                            componentLinks {
-                                id
-                                headComponentId
-                                tailComponentId
-
-                            }
-
-            }
-          }
-        }
-            """
-
-    def __init__(self, workflow_id: int, dataset_id: int, name: str,
-                 source_column_id: int, after_component_id: int = None, labelset_column_id: int = None,
-                 new_labelset_args: NewLabelsetArguments = None,
-                 new_questionnaire_args: NewQuestionaireArguments = None, model_training_options: str = None):
-        if labelset_column_id is not None and new_labelset_args is not None:
-            raise IndicoInputError("Cannot define both labelset_column_id and new_labelset_args, must be one "
-                                   "or the other.")
-        if labelset_column_id is None and new_labelset_args is None:
-            raise IndicoInputError("Must define one of either labelset_column_id or new_labelset_args.")
-
-        super().__init__(
-            self.query,
-            variables={
-                "workflowId": workflow_id,
-                "name": name,
-                "datasetId": dataset_id,
-                "sourceColumnId": source_column_id,
-                "labelsetColumnId": labelset_column_id,
-                "afterComponentId": after_component_id,
-                "modelTrainingOptions": model_training_options,
-                "newLabelsetArgs": self.__labelset_to_json(
-                    new_labelset_args) if new_labelset_args is not None else None,
-                "questionnaireArgs": self.__questionnaire_to_json(
-                    new_questionnaire_args) if new_questionnaire_args is not None else None
-
-            }
-        )
-
-    def __labelset_to_json(self, labelset: NewLabelsetArguments):
-        return {
-            "name": labelset.name,
-            "numLabelersRequired": labelset.num_labelers_required,
-            "datacolumnId": labelset.datacolumn_id,
-            "taskType": labelset.task_type,
-            "targetNames": labelset.target_names
-        }
-
-    def __questionnaire_to_json(self, questionnaire: NewQuestionaireArguments):
-        return {
-            "instructions": questionnaire.instructions,
-            "forceTextMode": questionnaire.force_text_mode,
-            "showPredictions": questionnaire.show_predictions,
-            "users": questionnaire.users
-
-        }
-
-    def process_response(self, response) -> Workflow:
-        return Workflow(**super().process_response(response)["addModelGroupComponent"]["workflow"])
